Reorganize ScalaPy completion into own package
