--- amm/compiler/src/main/scala-2.13/ammonite/compiler/ScalaPyCompletion.scala
@@ -1,3 +1,0 @@
-package ammonite.compiler
-
-trait ScalaPyCompletion extends ScalaPyCompletionBase
