+++ amm/compiler/src/main/scala-2.13/ammonite/compiler/scalapy/BasicCompletion.scala
@@ -1,0 +1,3 @@
+package ammonite.compiler.scalapy
+
+trait BasicCompletion extends BasicCompletionBase
