+++ amm/compiler/src/main/scala-2/ammonite/compiler/scalapy/Utils.scala
@@ -1,0 +1,7 @@
+package ammonite.compiler.scalapy
+
+import ammonite.compiler.Completion
+
+object Utils {
+
+}
